requirements.txt
flask==3.0.0
google-generativeai==0.4.1
langchain==0.1.0
langchain-google-genai==1.0.1
langchain-community==0.0.10
chromadb==0.4.22
pypdf2==3.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.1.0
transformers==4.36.0
sinling==0.3.6
transliterate==1.10.2
unidecode==1.3.7
streamlit==1.29.0
pandas==2.1.4
plotly==5.17.0
huggingface_hub==0.25.2